"Refactor benchmark tests for CPU usage and update dependencies"
Implemented changes to the benchmark tests in src/cpu/ios_macos.rs and transfered them to a new file named ios_macos.rs under the benchmarks/cpu directory. This helps in separating the CPU usage tests from the program file and adhere to the single-responsibility principle. Furthermore, added new dependency 'criterion' in Cargo.toml for benchmarking along with a corresponding bench setup. The removed lines in lib.rs is to allow these modifications, eliminating unnecessary features and dependencies.
