+++ Cargo.toml
@@ -31,3 +31,11 @@ mach =  "0.3"
 [build-dependencies]
 bindgen = "0.59"
 cc = "1.0"
+
+#[[bench]]
+#name = "cpu_ios_macos"
+#path = "benches/cpu/ios_macos.rs"
+#harness = false
+
+[dev-dependencies]
+criterion = "0.5.1"
